feat(migrations): Add SegundaMigration and RetiradaColunaEndereco migrations
- Created SegundaMigration to rename and restructure tables: Alerta to t_nimbus_alerta, Endereco to t_nimbus_endereco, and added t_nimbus_previsao table.
- Updated foreign key relationships and indexes accordingly.
- Created RetiradaColunaEndereco migration to remove unnecessary columns from t_nimbus_usuario and adjust foreign key constraints.
- Added new models: ObjetoAddUser, PreverCorpo, and RespostaPrever for handling user input and weather prediction responses.

+++ NimbusApi/Models/Previsao.cs
@@ -8,17 +8,17 @@ namespace NimbusApi.Models
 
         public DateTime DataHora { get; set; }
 
-        public int Temperatura { get; set; }
+        public float Temperatura { get; set; }
 
-        public int Chuva { get; set; }
+        public float Chuva { get; set; }
         
-        public int CodigoChuva { get; set; } 
+        public float CodigoChuva { get; set; } 
 
-        public int VelocidadeVento { get; set; }
+        public float VelocidadeVento { get; set; }
 
-        public int RajadaVento { get; set; }
+        public float RajadaVento { get; set; }
 
-        public int Umidade { get; set; }
+        public float Umidade { get; set; }
 
         public int IdBairro { get; set; }
